Removes dead code, unused containers, and relocates layout components into a logical dir. Also drops the pages stories
