+++ source/apps/shared/outdated-browser.tsx
@@ -2,16 +2,12 @@ import debug from 'debug';
 import React from 'react';
 import NoSSR from 'react-no-ssr';
 import { environment } from '../../environment';
+import OutdatedBrowser from '../../features/widgets/outdated-browser/components/OutdatedBrowser';
+import { Footer, Layout } from '../../layout';
 
 if (environment.DEBUG) {
   debug.enable(environment.DEBUG);
 }
-
-const Layout = require('../../layout/Layout').default;
-const OutdatedBrowser = require('../../features/widgets/outdated-browser/components/OutdatedBrowser')
-  .default;
-const FooterContainer = require('../../features/widgets/footer/containers/FooterContainer')
-  .default;
 
 export const OutdatedBrowserPage = () => (
   <NoSSR>
@@ -20,7 +16,7 @@ export const OutdatedBrowserPage = () => (
         text="Your web browser is out of date"
         updateBrowserUrl="http://browser-update.org/update.html?force_outdated=true"
       />
-      <FooterContainer />
+      <Footer />
     </Layout>
   </NoSSR>
 );
