--- source/features/blocks/containers/BlocksContainer.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { BlocksComponentContainer } from './BlocksComponentContainer';
-
-const BlocksContainer = () => <BlocksComponentContainer />;
-
-export default BlocksContainer;
