+++ source/features/blocks/containers/SearchBlockComponentContainer.tsx
@@ -4,7 +4,7 @@ import Search from '../../widgets/search/components/Search';
 import { useBlocks } from '../hooks';
 
 export const SearchBlockComponentContainerRaw = () => {
-  const { actions, store } = useBlocks();
+  const { actions } = useBlocks();
   return <Search onSearch={id => actions.searchBlockById.trigger({ id })} />;
 };
