+++ source/features/stake-pools/components/StakePoolsList.tsx
@@ -1,4 +1,3 @@
-import classnames from 'classnames';
 import { observer } from 'mobx-react-lite';
 import React, { FC, MouseEvent, useState } from 'react';
 import { getColorFromRange } from '../../../utils/colors';
