--- source/features/widgets/footer/containers/FooterComponentContainer.tsx
@@ -1,9 +1,0 @@
-import { observer } from 'mobx-react-lite';
-import React from 'react';
-import Footer from '../components/Footer';
-
-export const FooterComponentContainerRaw = () => {
-  return <Footer propFirst={''} propSecond={''} />;
-};
-
-export const FooterComponentContainer = observer(FooterComponentContainerRaw);
