--- source/features/widgets/footer/containers/FooterContainer.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { FooterComponentContainer } from './FooterComponentContainer';
-
-const FooterContainer = () => <FooterComponentContainer />;
-
-export default FooterContainer;
