--- source/features/widgets/header/containers/HeaderContainer.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { HeaderComponentContainer } from './HeaderComponentContainer';
-
-const HeaderContainer = () => <HeaderComponentContainer />;
-
-export default HeaderContainer;
