+++ source/layout/Footer.scss
