+++ source/layout/Header.tsx
@@ -2,11 +2,11 @@ import cx from 'classnames';
 import { observer } from 'mobx-react-lite';
 import Link from 'next/link';
 import React from 'react';
-import { BrandType } from '../../../../common/constants';
-import Search, { ISearchProps } from '../../search/components/Search';
+import { BrandType } from '../common/constants';
+import Search, { ISearchProps } from '../features/widgets/search/components/Search';
 import styles from './Header.scss';
 
-const CardanoLogo = require('../../../../static/assets/images/header/cardano-logo.svg');
+const CardanoLogo = require('../static/assets/images/header/cardano-logo.svg');
 
 export interface IHeaderProps {
   brandType?: BrandType;
@@ -15,7 +15,7 @@ export interface IHeaderProps {
   withSearch?: boolean;
 }
 
-const Header = (props: IHeaderProps) => {
+export const Header = observer((props: IHeaderProps) => {
   const { brandType, withSearch, searchProps } = props;
   const brandTypeStyle =
     brandType === BrandType.ENLARGED
@@ -69,11 +69,10 @@ const Header = (props: IHeaderProps) => {
       </div>
     </header>
   );
-};
+});
 
 Header.defaultProps = {
   brandType: BrandType.ENLARGED,
   withSearch: true,
 };
 
-export default observer(Header);
