+++ stories/layout.stories.tsx
@@ -2,10 +2,10 @@ import { storiesOf } from '@storybook/react';
 import { noop } from 'lodash';
 import React from 'react';
 import { BrandType } from '../source/common/constants';
-import { Header } from '../source/layout';
+import { Footer, Header } from '../source/layout';
 import { PaddingDecorator } from './support/PaddingDecorator';
 
-storiesOf('Header', module)
+storiesOf('Layout|Header', module)
   .addDecorator(story => <PaddingDecorator>{story()}</PaddingDecorator>)
   .add('Header', () => (
     <Header
@@ -14,3 +14,7 @@ storiesOf('Header', module)
       searchProps={{ onSearch: noop }}
     />
   ));
+
+storiesOf('Layout|Footer', module)
+  .addDecorator(story => <PaddingDecorator>{story()}</PaddingDecorator>)
+  .add('Footer', () => <Footer />);
